Fixes shared bundle pool problem
Buckminster prints the following message when using a shared bundle
pool: "Bundle Pool: Unable to load runtime repository: No repository
found at file:/..."

This commit makes it possible to set a user defined bundle pool
property: "org.eclipse.buckminster.core.bundle.pool". This property
takes precedence if set. If not, the property "osgi.syspath" is
consulted. If set to an existing directory, then the parent of that
directory is considered to be the bundle pool.
When none of the properties are set, Buckminster falls back to the
default platform behavior and assumes that the platform install
location is equal to the bundle pool.

Closes https://bugs.eclipse.org/bugs/show_bug.cgi?id=380097